OutputPath changed to be more universal
The OutputPath property in the .vbproj file was relative to Aaron's project path on his disk. Changed that to look for \Program Files (x86)\…, so to find the Program Files folder on the drive’s root. This should be more universal; still the condition is that the project directory is located on the system drive (usually “C:”)